Final Project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.31.0
google-generativeai==0.8.3
